changed name of diretory
