+++ KSkyBandQuery-C/KSkyBandQuery-C.xcodeproj/project.xcworkspace/xcshareddata/KSkyBandQuery-C.xccheckout
@@ -1,0 +1,41 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>IDESourceControlProjectFavoriteDictionaryKey</key>
+	<false/>
+	<key>IDESourceControlProjectIdentifier</key>
+	<string>B83BBCEF-09E3-4B8C-A6E1-450862B4A9CC</string>
+	<key>IDESourceControlProjectName</key>
+	<string>KSkyBandQuery-C</string>
+	<key>IDESourceControlProjectOriginsDictionary</key>
+	<dict>
+		<key>A5DE202BE512EC062FEED0D4D5C5761E7749CDE2</key>
+		<string>github.com:Armour/KeySkyBandQuery.git</string>
+	</dict>
+	<key>IDESourceControlProjectPath</key>
+	<string>KSkyBandQuery-C/KSkyBandQuery-C.xcodeproj</string>
+	<key>IDESourceControlProjectRelativeInstallPathDictionary</key>
+	<dict>
+		<key>A5DE202BE512EC062FEED0D4D5C5761E7749CDE2</key>
+		<string>../../..</string>
+	</dict>
+	<key>IDESourceControlProjectURL</key>
+	<string>github.com:Armour/KeySkyBandQuery.git</string>
+	<key>IDESourceControlProjectVersion</key>
+	<integer>111</integer>
+	<key>IDESourceControlProjectWCCIdentifier</key>
+	<string>A5DE202BE512EC062FEED0D4D5C5761E7749CDE2</string>
+	<key>IDESourceControlProjectWCConfigurations</key>
+	<array>
+		<dict>
+			<key>IDESourceControlRepositoryExtensionIdentifierKey</key>
+			<string>public.vcs.git</string>
+			<key>IDESourceControlWCCIdentifierKey</key>
+			<string>A5DE202BE512EC062FEED0D4D5C5761E7749CDE2</string>
+			<key>IDESourceControlWCCName</key>
+			<string>KSkyBandQuery</string>
+		</dict>
+	</array>
+</dict>
+</plist>
