+++ KSkyBandQuery-C/KSkyBandQuery-C/Bucket.h
