+++ KSkyBandQuery-C/KSkyBandQuery-C/HashTable.h
