+++ KSkyBandQuery-Cplus/KSkyBandQuery-CPlus/gtBucket.cpp
