+++ KSkyBandQuery-Cplus/KSkyBandQuery-CPlus/gtBucket.h
