--- gtKeySkyBandQueryPlus.xcodeproj/project.xcworkspace/xcshareddata/gtKeySkyBandQueryPlus.xccheckout
@@ -1,41 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-	<key>IDESourceControlProjectFavoriteDictionaryKey</key>
-	<false/>
-	<key>IDESourceControlProjectIdentifier</key>
-	<string>D0AB8E61-539A-4DBA-AD09-00CD90F16876</string>
-	<key>IDESourceControlProjectName</key>
-	<string>gtKeySkyBandQueryPlus</string>
-	<key>IDESourceControlProjectOriginsDictionary</key>
-	<dict>
-		<key>A5DE202BE512EC062FEED0D4D5C5761E7749CDE2</key>
-		<string>github.com:Armour/KeySkyBandQuery.git</string>
-	</dict>
-	<key>IDESourceControlProjectPath</key>
-	<string>gtKeySkyBandQueryPlus.xcodeproj</string>
-	<key>IDESourceControlProjectRelativeInstallPathDictionary</key>
-	<dict>
-		<key>A5DE202BE512EC062FEED0D4D5C5761E7749CDE2</key>
-		<string>../..</string>
-	</dict>
-	<key>IDESourceControlProjectURL</key>
-	<string>github.com:Armour/KeySkyBandQuery.git</string>
-	<key>IDESourceControlProjectVersion</key>
-	<integer>111</integer>
-	<key>IDESourceControlProjectWCCIdentifier</key>
-	<string>A5DE202BE512EC062FEED0D4D5C5761E7749CDE2</string>
-	<key>IDESourceControlProjectWCConfigurations</key>
-	<array>
-		<dict>
-			<key>IDESourceControlRepositoryExtensionIdentifierKey</key>
-			<string>public.vcs.git</string>
-			<key>IDESourceControlWCCIdentifierKey</key>
-			<string>A5DE202BE512EC062FEED0D4D5C5761E7749CDE2</string>
-			<key>IDESourceControlWCCName</key>
-			<string>KeySkyBandQuery</string>
-		</dict>
-	</array>
-</dict>
-</plist>
